% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEffect-ale.R
\name{calculate.ale.cat}
\alias{calculate.ale.cat}
\title{Compute ALE for 1 categorical feature}
\usage{
calculate.ale.cat(dat, run.prediction, feature.name)
}
\arguments{
\item{dat}{the data.frame with same columns as training data}

\item{run.prediction}{Predict function of type: f(newdata)}

\item{feature.name}{The column name of the feature for which to compute ALE}
}
\description{
Compute ALE for 1 categorical feature
}
