% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_counterfactuals.R
\name{compute_diversity}
\alias{compute_diversity}
\title{Calculate diversity}
\usage{
compute_diversity(df, range)
}
\value{
(numeric(1))
}
\description{
The diversity is equal to pair-wise mean of Gower distances.
}
\section{Arguments}{
 
\describe{
\item{df: }{(data.frame)\cr Pareto set.}
\item{range: }{(numeric)\cr Vector of ranges for numeric features. 
Must have same ordering as columns in df.}
}
}

