% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{extract.glment.effects}
\alias{extract.glment.effects}
\title{Extract glmnet effects}
\usage{
extract.glment.effects(betas, best.index, x.scaled, x.original)
}
\arguments{
\item{betas}{glment$beta
Assuming that the first row is the x.interest}
}
\description{
Extract glmnet effects
}
