% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditional.R
\name{fit_conditionals}
\alias{fit_conditionals}
\title{Fit conditional models}
\usage{
fit_conditionals(data, ctrl = ctree_control())
}
\arguments{
\item{data}{data.frame with data for which to fit the conditional models}
}
\value{
list of Conditional R6 objects
}
\description{
Needed for conditional PDP and Feature Importance.
}
