% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_counterfactuals.R
\name{get_diverse_solutions}
\alias{get_diverse_solutions}
\title{Algorithm to remove solutions of final set}
\usage{
get_diverse_solutions(fitness, pareto.set, nr.solutions)
}
\description{
Algorithm to remove solutions of final set
}
\section{Arguments}{
 
\describe{
\item{fitness: }{(data.frame)\cr Data frame of fitness values. Each column 
represents one objective.}
\item{pareto.set: }{(data.frame)\cr Corresponding Pareto set to 
fitness.}
\item{nr.solutions: }{(numeric(1))\cr Number of solutions that should
be returned.} 
}
}

