% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has.predict}
\alias{has.predict}
\title{returns TRUE if object has predict function}
\usage{
has.predict(object)
}
\arguments{
\item{object}{The object to check.}
}
\description{
returns TRUE if object has predict function
}
