% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iml.R
\docType{package}
\name{iml-package}
\alias{iml}
\alias{iml-package}
\title{Make machine learning models and predictions interpretable}
\description{
The iml package provides tools to analyse machine learning models and predictions.
}
\seealso{
\href{https://christophm.github.io/interpretable-ml-book/agnostic}{Book on Interpretable Machine Learning}
}
\author{
\strong{Maintainer}: Christoph Molnar \email{christoph.molnar@gmail.com}

}
