% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEffect-ale.R
\name{impute_cells}
\alias{impute_cells}
\title{Impute missing cells of grid}
\usage{
impute_cells(cell.dat, grid1 = NULL, grid2, x1.ind = 1, x2.ind = 2)
}
\arguments{
\item{cell.dat}{data.table with at least 4 columns: .yhat.diff and the two interval indices.
Make sure that empty cells are also included and cell.dat is not the sparse representation.}

\item{grid1}{data.frame where each row is the actual value for a given interval index for feature 1.
If empty impute_cells  assumes that the feature is categorical (factor).}

\item{grid2}{data.frame where each row is the actual value for a given interval index for feature 2}

\item{x1.ind}{column number or name of cell.dat for feature 1. If one feature is categorical, has to be x1}

\item{x2.ind}{column number or name of cell.dat for feature 2}
}
\description{
by default assumes first column of cell.dat is x1 and second is x2
leave grid1 NULL if feature x1 is a factor
the difference variable has to be named .yhat.diff
}
\keyword{internal}
