% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEffect-ale.R
\name{impute_intervals}
\alias{impute_intervals}
\title{Impute intervals with no data inside}
\usage{
impute_intervals(interval.dat, x.col)
}
\arguments{
\item{interval.dat}{data.frame with column .yhat.diff and interval index and interval position}

\item{x.col}{The index of the feature column}
}
\description{
Should only occur when type = equidistant was used
}
