% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interaction.R
\name{Interaction}
\alias{Interaction}
\title{Interaction strength}
\format{\code{\link{R6Class}} object.}
\description{
\code{TreeSurrogate} measures the interaction strength of features in a prediction model.
}
\section{Usage}{

\preformatted{
tree = Interaction$new(predictor, maxdepth = 2, tree.args = NULL, run = TRUE)

plot(tree)
predict(tree, newdata)
tree$results
print(tree)
}
}

\section{Arguments}{


For Interaction$new():
\describe{
\item{predictor: }{(Predictor)\cr 
The object (created with Predictor$new()) holding the machine learning model and the data.}
\item{maxdepth: }{(`numeric(1)`)\cr The maximum depth of the tree. Default is 2.}
\item{run: }{(`logical(1)`)\cr Should the Interpretation method be run?}
\item{tree.args: }{(named list)\cr Further arguments for \code{ctree}.}
}
}

\section{Details}{
  
A conditional inference tree is fitted on the predicted \eqn{\hat{y}} from the machine learning model and the data.
The \code{partykit} package and function are used to fit the tree. 
By default a tree of maximum depth of 2 is fitted to improve interpretability.
}

\section{Fields}{

\describe{
\item{maxdepth: }{(`numeric(1)`)\cr The maximum tree depth.}
\item{predictor: }{(Predictor)\cr The prediction model that was analysed.}
\item{r.squared}{(`numeric(1|n.classes)`)\cr R squared measures how well the decision tree approximates the underlying model. 
It is calculated as 1 - (variance of prediction differences / variance of black box model predictions).
For the multi-class case, r.squared contains one measure per class.}
\item{results: }{(data.frame)\cr Data.frame with sampled feature X together with the leaf node information (columns .node and .path) 
and the predicted \eqn{\hat{y}} for tree and machine learning model (columns starting with .y.hat).}
\item{tree: }{(party)\cr The fitted tree. See also \link[partykit]{ctree}.}
}
}

\section{Methods}{

\describe{
\item{plot()}{method to plot the leaf nodes of the surrogate decision tree. See \link{plot.TreeSurrogate}.}
\item{predict()}{method to predict new data with the tree. See also \link{predict.TreeSurrogate}}
\item{\code{run()}}{[internal] method to run the interpretability method. Use \code{obj$run(force = TRUE)} to force a rerun.}
\item{\code{clone()}}{[internal] method to clone the R6 object.}
\item{\code{initialize()}}{[internal] method to initialize the R6 object.}
}
}

\references{
Craven, M., & Shavlik, J. W. (1996). Extracting tree-structured representations of trained networks. In Advances in neural information processing systems (pp. 24-30).
}
\seealso{
\link{Partial}
}
