% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lime.r
\name{lime}
\alias{lime}
\title{Local models}
\usage{
lime(object, X, sample.size = 100, k = 3, x.interest, class = NULL, ...)
}
\arguments{
\item{object}{The object is the machine learning model. Different types are allowed. 
Recommended are mlr WrappedModel and caret train objects. The \code{object} can also be 
a function that predicts the outcome given features or anything with an S3 predict function,
like an object from class \code{lm}.}

\item{X}{data.frame with the data for the prediction model}

\item{class}{In case of classification, class specifies the class for which to predict the probability. 
By default the multiclass classification is done.}
}
\description{
Local models
}
