% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictor.R
\name{makePredictor}
\alias{makePredictor}
\title{Make a Predictor() object}
\usage{
makePredictor(object, class = NULL, predict.args = NULL, ...)
}
\arguments{
\item{object}{function, mlr WrappedObject, S3 class with predict function, or caret train object}

\item{class}{In case of classification, class specifies the class for which to predict the probability. 
By default the first class in the prediction (first column) is chosen.}

\item{predict.args}{named list with arguments passed down to the prediction model}

\item{...}{further arguments for the prediction functions}
}
\value{
object of type Predictor
}
\description{
Make a Predictor() object
}
