% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalEffects.R
\name{MarginalEffects}
\alias{MarginalEffects}
\title{Marginal effects of features on the prediction}
\format{\code{\link{R6Class}} object.}
\description{
\code{MarginalEffects} computes and plots marginal feature effects of prediction models.
}
\section{Usage}{

\preformatted{
effect = FeatureEffect$new(predictor, feature, step.size,  method = "forward", 
    grid.size = 4)

plot(effect)
effect$results
effects$ame
print(effect)
}
}

\section{Arguments}{


For MarginalEffect$new():
\describe{
\item{predictor: }{(Predictor)\cr 
The object (created with Predictor$new()) holding the machine learning model and the data.}
\item{feature: }{(`character(n)` | ``numeric(n)) \cr The feature names or indices for which to compute the effects.}
\item{method: }{(`character(1)`)\cr 
'forward' for absolute forward step marginal effect, 
'derivative' for the right sided finite difference, 
}
\item{step.size: }{(`numeric(n)`)\cr The size of the forward steps. Only applicable if method='forward'}
\item{eps: }{`numeric(n)`)\cr The finite differences step size. Only applicable if method='derivative'}
}
}

\section{Details}{


TODO
}

\section{Methods}{

\describe{
\item{plot()}{method to plot the marginal effects. See \link{plot.MarginalEffects}}
\item{\code{clone()}}{[internal] method to clone the R6 object.}
\item{\code{initialize()}}{[internal] method to initialize the R6 object.}
}
}

\examples{
# TODO: Update example here
# We train a random forest on the Boston dataset:
if (require("randomForest")) {
data("Boston", package  = "MASS")
rf = randomForest(medv ~ ., data = Boston, ntree = 50)
mod = Predictor$new(rf, data = Boston)

# Compute the accumulated local effects for the first feature
eff = MarginalEffects$new(mod, feature = "rm", step.size = 1)
eff$plot()
eff$ame

# Again, but this time with derivative method 
eff = MarginalEffects$new(mod, feature = "rm", method = "derivative")
plot(eff)
eff$ame

# MarginalEffects for multiple  features: 
eff = MarginalEffects$new(mod, feature = c("lstat", "rm"), step.size = c(1, 0.5))
eff$plot()   

# Multiclass classification
rf = randomForest(Species ~ ., data = iris, ntree=50)
mod = Predictor$new(rf, data = iris, type = "prob")
eff = MarginalEffects$new(mod, feature = "Petal.Width", step.size = 1)
eff$plot()
eff$ame
}
}
\references{
TODO: reference paper here
}
\seealso{
\link{plot.MarginalEffects}
}
