% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Model}
\alias{Model}
\title{Prediction Model object}
\format{\code{\link{R6Class}} object.}
\value{
object of type Model
}
\description{
A \code{Model} object wraps any machine learning model (mlr, caret, randomForest, ...). 
The object is to be used for the interpretation methods.
}
\section{Usage}{

\preformatted{
model = Model$new(object, class=NULL, predict.args = NULL)

model$predict(newdata)
}
}

\section{Details}{
 
In case of classification, the model must return the probabilities. Model doesn't handle machine learning
models that return the labels.
}

\section{Fields}{

\describe{
\item{class}{The class column to be returned.}
\item{prediction.colnames}{The column names of the predictions.}
\item{task}{The inferred prediction task: "classification" or "regression".}
}
}

\section{Methods}{

\describe{
\item{predict(newdata)}{method to predict new data with the machine learning model. See also \link{predict.TreeSurrogate}}
\item{\code{clone()}}{[internal] method to clone the R6 object.}
\item{\code{initialize()}}{[internal] method to initialize the R6 object.}
}
}

