% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_levels}
\alias{order_levels}
\title{Order levels of a categorical features}
\usage{
order_levels(dat, feature.name)
}
\arguments{
\item{dat}{data.frame with the training data}

\item{feature.name}{the name of the categorical feature}
}
\value{
the order of the levels (not levels itself)
}
\description{
Orders the levels by their similarity in other features.
Computes per feature the distance, sums up all distances and does multi-dimensional scaling
}
