% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureImp.R
\name{plot.FeatureImp}
\alias{plot.FeatureImp}
\title{Feature importance plot}
\usage{
\method{plot}{FeatureImp}(x, sort = TRUE, ...)
}
\arguments{
\item{x}{The feature importance. An FeatureImp R6 object}

\item{sort}{logical. Should the features be sorted in descending order? Defaults to TRUE.}

\item{...}{Further arguments for the objects plot function}
}
\value{
ggplot2 plot object
}
\description{
plot.FeatureImp() plots the feature importance results of an FeatureImp object.
}
\details{
For examples see \link{makeFeatureImp}
}
\seealso{
\link{makeFeatureImp}
}
