% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.r
\name{plot.ICE}
\alias{plot.ICE}
\title{Individual conditional expectation plots}
\usage{
\method{plot}{ICE}(x, ...)
}
\arguments{
\item{x}{The individual conditional expectation curves. An ICE R6 object}

\item{...}{Further arguments for the objects plot function}
}
\value{
ggplot2 plot object
}
\description{
plot.ICE() plots individiual expectation curves for each observation for one feature.
}
\details{
For examples see \link{ice}
}
\seealso{
\link{ice}
}
