% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{plot.Importance}
\alias{plot.Importance}
\title{Feature importance plot}
\usage{
\method{plot}{Importance}(object, sort = TRUE)
}
\arguments{
\item{object}{The feature importance. An Importance R6 object}

\item{sort}{logical. Should the features be sorted in descending order? Defaults to TRUE.}
}
\value{
ggplot2 plot object
}
\description{
plot.Importance() plots the feature importance results of an Importance object.
}
\details{
For examples see \link{importance}
}
\seealso{
\link{importance}
}
