% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interaction.R
\name{plot.Interaction}
\alias{plot.Interaction}
\title{Plot Interaction}
\usage{
\method{plot}{Interaction}(x)
}
\arguments{
\item{x}{An Interaction R6 object}
}
\value{
ggplot2 plot object
}
\description{
plot.Interaction() plots the results of an Interaction object.
}
\examples{
# We train a random forest on the Boston dataset:
if (require("randomForest")) {
data("Boston", package  = "MASS")
rf = randomForest(medv ~ ., data = Boston, ntree = 50)
mod = Predictor$new(rf, data = Boston)

# Compute the interactions
pdp.obj = Interaction$new(mod)

# Plot the results directly
plot(pdp.obj)
}
}
\seealso{
\link{Interaction}
}
