% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lime.R
\name{plot.Lime}
\alias{plot.Lime}
\title{Lime plot}
\usage{
\method{plot}{Lime}(object)
}
\arguments{
\item{object}{A Lime R6 object}
}
\value{
ggplot2 plot object
}
\description{
plot.Lime() plots the feature effects of the Lime model.
}
\details{
For examples see \link{makeLime}
}
\seealso{
\link{makeLime}
}
