% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalModel.R
\name{plot.LocalModel}
\alias{plot.LocalModel}
\title{Plot Local Model}
\usage{
\method{plot}{LocalModel}(object)
}
\arguments{
\item{object}{A LocalModel R6 object}
}
\value{
ggplot2 plot object
}
\description{
plot.LocalModel() plots the feature effects of a LocalModel object.
}
\details{
For examples see \link{LocalModel}
}
\seealso{
\link{LocalModel}
}
