% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalEffects.R
\name{plot.MarginalEffects}
\alias{plot.MarginalEffects}
\title{Plot MarginalEffects}
\usage{
\method{plot}{MarginalEffects}(x)
}
\arguments{
\item{x}{A MarginalEffects R6 object}
}
\value{
ggplot2 plot object
}
\description{
plot.MarginalEffects() plots the results of a MarginalEffects object.
}
\examples{
# We train a random forest on the Boston dataset:
if (require("randomForest")) {
data("Boston", package  = "MASS")
rf = randomForest(medv ~ ., data = Boston, ntree = 50)
mod = Predictor$new(rf, data = Boston)

# Compute the marginal effects for the first feature
eff = MarginalEffects$new(mod, feature = "crim", step.size = 1)

# Plot the results directly
plot(eff)
}
}
\seealso{
\link{MarginalEffects}
}
