% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialDependence.R
\name{plot.PartialDependence}
\alias{plot.PartialDependence}
\title{Plot Partial Dependence}
\usage{
\method{plot}{PartialDependence}(x)
}
\arguments{
\item{x}{A PartialDependence R6 object}
}
\value{
ggplot2 plot object
}
\description{
plot.PartialDependence() plots the results of a PartialDependence object.
}
\details{
For examples see \link{PartialDependence}
}
\seealso{
\link{PartialDependence}
}
