% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.r, R/pdp.r
\name{plot.PDP}
\alias{plot.PDP}
\alias{plot.PDP}
\title{Individual conditional expectation plots}
\usage{
\method{plot}{PDP}(x, ...)

\method{plot}{PDP}(x, ...)
}
\arguments{
\item{x}{The partial dependence. A PDP R6 object}

\item{...}{Further arguments for the objects plot function}

\item{object}{The individual conditional expectation curves. An ICE R6 object}
}
\value{
ggplot2 plot object

ggplot2 plot object
}
\description{
plot.ICE() plots individiual expectation curves for each observation for one feature.

plot.PDP() plots a line for a single feature and a tile plot for 2 features.
}
\details{
For examples see \link{ice}

For examples see \link{pdp}
}
\seealso{
\link{ice}

\link{pdp}
}
