% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapley.R
\name{plot.Shapley}
\alias{plot.Shapley}
\title{Shapley plot}
\usage{
\method{plot}{Shapley}(object)
}
\arguments{
\item{object}{A Shapley R6 object}
}
\value{
ggplot2 plot object
}
\description{
plot.Shapley() plots the Shapley values - the contributions of feature values to the prediction.
}
\details{
For examples see \link{makeShapley}
}
\seealso{
\link{makeShapley}
}
