% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSurrogate.R
\name{plot.TreeSurrogate}
\alias{plot.TreeSurrogate}
\title{Plot Tree Surrogate}
\usage{
\method{plot}{TreeSurrogate}(object)
}
\arguments{
\item{object}{A TreeSurrogate R6 object}
}
\value{
ggplot2 plot object
}
\description{
Plot the response for newdata of a TreeSurrogate object.
Each plot facet is one leaf node and visualises the distribution of the \eqn{\hat{y}}
from the machine learning model.
}
\details{
For examples see \link{TreeSurrogate}
}
\seealso{
\link{TreeSurrogate}
}
