% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lime.R
\name{predict.LIME}
\alias{predict.LIME}
\title{LIME prediction}
\usage{
\method{predict}{LIME}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A LIME R6 object}

\item{newdata}{A data.frame for which to predict}

\item{...}{Further arguments for the objects predict function}
}
\value{
A data.frame with the predicted outcome.
}
\description{
Predict the response for newdata with the LIME model.
}
\details{
This function makes the LIME object call 
its iternal object$predict() method. 
For examples see \link{lime}
}
\seealso{
\link{lime}
}
