% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{probs.to.labels}
\alias{probs.to.labels}
\title{Turn class probabilities into class labels}
\usage{
probs.to.labels(prediction)
}
\arguments{
\item{prediction}{Prediction object.}
}
\description{
Turn class probabilities into class labels
}
