% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEffect-ale.R
\name{calculate.ale.num.cat}
\alias{calculate.ale.num.cat}
\title{Compute ALE for 2 features, one numerical, one categorical}
\usage{
calculate.ale.num.cat(dat, run.prediction, feature.name, grid.size)
}
\arguments{
\item{dat}{the data.frame with same columns as training data}

\item{run.prediction}{Predict function of type: f(newdata)}

\item{feature.name}{The column name of the features for which to compute ALE}

\item{grid.size}{The number of intervals for the numerical feature}
}
\description{
Compute ALE for 2 features, one numerical, one categorical
}
