% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEffect-ale.R
\name{calculate.ale.num.num}
\alias{calculate.ale.num.num}
\title{Compute ALE for 2 numerical features}
\usage{
calculate.ale.num.num(dat, run.prediction, feature.name, grid.size)
}
\arguments{
\item{dat}{the data.frame with same columns as training data}

\item{run.prediction}{Predict function of type: f(newdata)}

\item{feature.name}{The column names of the feature for which to compute ALE}

\item{grid.size}{The number of cells}
}
\description{
Compute ALE for 2 numerical features
}
