% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEffect-ale.R
\name{calculate.ale.num}
\alias{calculate.ale.num}
\title{Compute ALE for 1 numerical feature}
\usage{
calculate.ale.num(dat, run.prediction, feature.name, grid.dt)
}
\arguments{
\item{dat}{the data.frame with same columns as training data}

\item{run.prediction}{Predict function of type: f(newdata)}

\item{feature.name}{The column name of the feature for which to compute ALE}

\item{grid.dt}{data.table with single column with grid values for the numerical feature}
}
\description{
Compute ALE for 1 numerical feature
}
\keyword{internal}
