% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{extract.glmnet.effects}
\alias{extract.glmnet.effects}
\title{Extract glmnet effects}
\usage{
extract.glmnet.effects(betas, best.index, x.scaled, x.original)
}
\arguments{
\item{betas}{glmnet$beta
Assuming that the first row is the x.interest}
}
\description{
Extract glmnet effects
}
