% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract.glmnet.effects}
\alias{extract.glmnet.effects}
\title{Extract glmnet effects}
\usage{
extract.glmnet.effects(betas, best.index, x.recoded, x.original)
}
\arguments{
\item{betas}{glmnet$beta}

\item{best.index}{index k}

\item{x.recoded}{the recoded version of x}

\item{x.original}{the original x
Assuming that the first row is the x.interest}
}
\description{
Extract glmnet effects
}
