% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iml.R
\docType{package}
\name{iml-package}
\alias{iml}
\alias{iml-package}
\title{Make machine learning models and predictions interpretable}
\description{
The iml package provides tools to analyze machine learning models and predictions.
}
\seealso{
\href{https://christophm.github.io/interpretable-ml-book/agnostic}{Book on Interpretable Machine Learning}
}
\author{
\strong{Maintainer}: Christoph Molnar \email{christoph.molnar@gmail.com}

Authors:
\itemize{
  \item Patrick Schratz \email{patrick.schratz@gmail.com} (\href{https://orcid.org/0000-0003-0748-6624}{ORCID})
}

}
