% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpretationMethod.R
\name{InterpretationMethod}
\alias{InterpretationMethod}
\title{Interpretation Method}
\description{
Superclass container for Interpretation Method objects
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{results}}{\link{data.frame}\cr
The aggregated results of the experiment}

\item{\code{predictor}}{Predictor object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InterpretationMethod-new}{\code{InterpretationMethod$new()}}
\item \href{#method-InterpretationMethod-plot}{\code{InterpretationMethod$plot()}}
\item \href{#method-InterpretationMethod-print}{\code{InterpretationMethod$print()}}
\item \href{#method-InterpretationMethod-clone}{\code{InterpretationMethod$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InterpretationMethod-new"></a>}}
\if{latex}{\out{\hypertarget{method-InterpretationMethod-new}{}}}
\subsection{Method \code{new()}}{
Create an InterpretationMethod object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InterpretationMethod$new(predictor)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictor}}{\link{Predictor}\cr
The object (created with \code{Predictor$new()}) holding the machine
learning model and the data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InterpretationMethod-plot"></a>}}
\if{latex}{\out{\hypertarget{method-InterpretationMethod-plot}{}}}
\subsection{Method \code{plot()}}{
Plot function. Calls \code{private$generatePlot()} of the
respective subclass.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InterpretationMethod$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Passed to \code{private$generatePlot()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InterpretationMethod-print"></a>}}
\if{latex}{\out{\hypertarget{method-InterpretationMethod-print}{}}}
\subsection{Method \code{print()}}{
Printer for InterpretationMethod objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InterpretationMethod$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InterpretationMethod-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InterpretationMethod-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InterpretationMethod$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
