% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEffect.R
\name{Partial}
\alias{Partial}
\title{Effect of one or two feature(s) on the model predictions (deprecated)}
\format{\code{\link{R6Class}} object.}
\description{
Deprecated, please use 'FeatureEffect', see ?FeatureEffect
}
\seealso{
\link{FeatureEffect}
}
