% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ice.R
\name{plot.Ice}
\alias{plot.Ice}
\title{Individual conditional expectation plots}
\usage{
\method{plot}{Ice}(x, ...)
}
\arguments{
\item{x}{The individual conditional expectation curves. An Ice R6 object}

\item{...}{Further arguments for the objects plot function}
}
\value{
ggplot2 plot object
}
\description{
plot.Ice() plots individiual expectation curves for each observation for one feature.
}
\details{
For examples see \link{Ice}
}
\seealso{
\link{Ice}
}
