% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ice.R
\name{plot.Ice}
\alias{plot.Ice}
\title{Plot ICE (Individual Conditional Expectation)}
\usage{
\method{plot}{Ice}(x)
}
\arguments{
\item{x}{An Ice R6 object}
}
\value{
ggplot2 plot object
}
\description{
plot.Ice() plots the individiual expectation results from an Ice object.
}
\details{
For examples see \link{Ice}
}
\seealso{
\link{Ice}
}
