% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.imp.R
\name{plot.Importance}
\alias{plot.Importance}
\title{Feature importance plot}
\usage{
\method{plot}{Importance}(x, sort = TRUE, ...)
}
\arguments{
\item{x}{The feature importance. An Importance R6 object}

\item{sort}{logical. Should the features be sorted in descending order? Defaults to TRUE.}

\item{...}{Further arguments for the objects plot function}
}
\value{
ggplot2 plot object
}
\description{
plot.Importance() plots the feature importance results of an Importance object.
}
\details{
For examples see \link{feature.imp}
}
\seealso{
\link{feature.imp}
}
