% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lime.R
\name{plot.LIME}
\alias{plot.LIME}
\title{LIME plot}
\usage{
\method{plot}{LIME}(object)
}
\arguments{
\item{object}{A LIME R6 object}
}
\value{
ggplot2 plot object
}
\description{
plot.LIME() plots the feature effects of the LIME model.
}
\details{
For examples see \link{lime}
}
\seealso{
\link{lime}
}
