% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialDependence.R
\name{plot.PartialDependence}
\alias{plot.PartialDependence}
\title{Partial dependence plot}
\usage{
\method{plot}{PartialDependence}(x, ...)
}
\arguments{
\item{x}{The partial dependence. A PartialDependence R6 object}

\item{...}{Further arguments for the objects plot function}
}
\value{
ggplot2 plot object
}
\description{
plot.PartialDependence() plots a line for a single feature and a tile plot for 2 features.
}
\details{
For examples see \link{PartialDependence}
}
\seealso{
\link{PartialDependence}
}
