% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialDependence.R
\name{plot.PartialDependence}
\alias{plot.PartialDependence}
\title{Plot Partial Dependence}
\usage{
\method{plot}{PartialDependence}(x)
}
\arguments{
\item{x}{A PartialDependence R6 object}
}
\value{
ggplot2 plot object
}
\description{
plot.PartialDependence() plots the results of a PartialDependence object.
}
\examples{
# We train a random forest on the Boston dataset:
if (require("randomForest")) {
data("Boston", package  = "MASS")
rf = randomForest(medv ~ ., data = Boston, ntree = 50)
mod = Predictor$new(rf, data = Boston)

# Compute the partial dependence for the first feature
pdp.obj = PartialDependence$new(mod, feature = "crim")

# Plot the results directly
plot(pdp.obj)
}
}
\seealso{
\link{PartialDependence}
}
