% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp.r
\name{plot.PDP}
\alias{plot.PDP}
\title{Partial dependence plot}
\usage{
\method{plot}{PDP}(x, ...)
}
\arguments{
\item{x}{The partial dependence. A PDP R6 object}

\item{...}{Further arguments for the objects plot function}
}
\value{
ggplot2 plot object
}
\description{
plot.PDP() plots a line for a single feature and a tile plot for 2 features.
}
\details{
For examples see \link{pdp}
}
\seealso{
\link{pdp}
}
