% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapley.R
\name{plot.Shapley}
\alias{plot.Shapley}
\title{Plot Shapley}
\usage{
\method{plot}{Shapley}(object, sort = TRUE)
}
\arguments{
\item{object}{A Shapley R6 object}

\item{sort}{logical. Should the feature values be sorted by Shapley value? Ignored for multi.class output.}
}
\value{
ggplot2 plot object
}
\description{
plot.Shapley() plots the Shapley values - the contributions of feature values to the prediction.
}
\details{
For examples see \link{Shapley}
}
\seealso{
\link{Shapley}
}
