% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-surrogate.r
\name{plot.TreeSurrogate}
\alias{plot.TreeSurrogate}
\title{Surrogate tree visualisation}
\usage{
\method{plot}{TreeSurrogate}(object)
}
\arguments{
\item{object}{The surrogate tree. A TreeSurrogate R6 object}
}
\value{
ggplot2 plot object
}
\description{
Predict the response for newdata with the surrogate tree
Each plot facet is one leaf node and visualises the distribution of the \eqn{\hat{y}}
from the machine learning model. 
This makes the TreeSurrogate object call 
its iternal object$plot() method.
}
\details{
For examples see \link{tree.surrogate}
}
\seealso{
\link{tree.surrogate}
}
