% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lime.R
\name{predict.Lime}
\alias{predict.Lime}
\title{Lime prediction}
\usage{
\method{predict}{Lime}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A Lime R6 object}

\item{newdata}{A data.frame for which to predict}

\item{...}{Further arguments for the objects predict function}
}
\value{
A data.frame with the predicted outcome.
}
\description{
Predict the response for newdata with the Lime model.
}
\details{
This function makes the Lime object call 
its iternal object$predict() method. 
For examples see \link{Lime}
}
\seealso{
\link{Lime}
}
