% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalModel.R
\name{predict.LocalModel}
\alias{predict.LocalModel}
\title{Predict LocalModel}
\usage{
\method{predict}{LocalModel}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A LocalModel R6 object}

\item{newdata}{A data.frame for which to predict}

\item{...}{Further arguments for the objects predict function}
}
\value{
A data.frame with the predicted outcome.
}
\description{
Predict the response for newdata with the LocalModel model.
}
\details{
For examples see \link{LocalModel}
}
\seealso{
\link{LocalModel}
}
